+++ src/feature_flag/services/feature_flag_service.py
@@ -1,4 +1,8 @@
+from dataclasses import replace
 from typing import Optional
+from uuid import UUID
+
+from redis import RedisCluster
 
 from feature_flag.models.feature_flag import FeatureFlag
 from feature_flag.core.base_repository import BaseRepository
@@ -10,32 +14,35 @@ class FeatureFlagService:
         self.repository = repository
         self.cache = cache
 
-    def create_feature_flag(self, flag_data: dict):
-        entity_id = self.repository.insert(entity=FeatureFlag(**flag_data))
-        feature_flag = self.repository.get_by_id(
-            entity_id=entity_id, entity_class=FeatureFlag
+    async def create_feature_flag(self, flag_data: dict):
+        entity_id = await self.repository.insert(entity=FeatureFlag(**flag_data))
+        feature_flag = await self.repository.get_by_id(
+            entity_id=str(entity_id), entity_class=FeatureFlag
         )
+
+        if isinstance(feature_flag.id, UUID):
+            feature_flag.id = str(feature_flag.id)
 
         if self.cache:
             self.cache.set(key=feature_flag.id, value=feature_flag)
 
         return feature_flag
 
-    def get_feature_flag(self, entity_id: str) -> Optional[FeatureFlag]:
+    async def get_feature_flag(self, entity_id: str) -> Optional[FeatureFlag]:
         if self.cache:
             cached_flag = self.cache.get(key=entity_id)
             if cached_flag:
                 return cached_flag
 
-        flag = self.repository.get_by_id(entity_id=entity_id, entity_class=FeatureFlag)
+        flag = await self.repository.get_by_id(entity_id=entity_id, entity_class=FeatureFlag)
 
         if flag and self.cache:
             self.cache.set(key=entity_id, value=flag)
 
         return flag
 
-    def update_feature_flag(self, entity_id: str, flag_data: dict):
-        existing_flag = self.get_feature_flag(entity_id)
+    async def update_feature_flag(self, entity_id: str, flag_data: dict):
+        existing_flag = await self.get_feature_flag(entity_id)
         if not existing_flag:
             raise ValueError(f"Feature flag with ID {entity_id} does not exist.")
 
@@ -51,20 +58,20 @@ class FeatureFlagService:
             setattr(existing_flag, key, value)
 
         # Update the repository
-        self.repository.update(entity=existing_flag)
+        await self.repository.update(entity=existing_flag)
 
         if self.cache:
             self.cache.set(key=existing_flag.id, value=existing_flag)
 
-    def enable_feature_flag(self, entity_id: str):
+    async def enable_feature_flag(self, entity_id: str):
         flag_data = {"enabled": True}
-        self.update_feature_flag(entity_id, flag_data)
+        await self.update_feature_flag(entity_id, flag_data)
 
-    def disable_feature_flag(self, entity_id: str):
+    async def disable_feature_flag(self, entity_id: str):
         flag_data = {"enabled": False}
-        self.update_feature_flag(entity_id, flag_data)
+        await self.update_feature_flag(entity_id, flag_data)
 
-    def delete_feature_flag(self, entity_id: str):
-        self.repository.delete(entity_id=entity_id)
+    async def delete_feature_flag(self, entity_id: str):
+        await self.repository.delete(entity_id=entity_id)
         if self.cache:
             self.cache.delete(key=entity_id)