+++ tests/test_app.py
@@ -1,16 +1,18 @@
 from fastapi import FastAPI, Depends
 from fastapi.encoders import jsonable_encoder
+from redis import RedisCluster
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from feature_flag.core.cache import RedisCache
 from feature_flag.models.feature_flag import FeatureFlag
 from feature_flag.repositories.postgres_repository import PostgresRepository
 from feature_flag.services.feature_flag_service import FeatureFlagService
-from tests.test_utils import get_redis_connection, get_db_connection
+from tests.test_utils import get_redis_connection, get_db_session
 
 
 def get_feature_flag_service(
-    db_connection=Depends(get_db_connection),
-    redis_connection=Depends(get_redis_connection),
+    db_connection: AsyncSession = Depends(get_db_session),
+    redis_connection: RedisCluster = Depends(get_redis_connection),
 ):
     repository = PostgresRepository(db_connection)
     cache = RedisCache(redis_connection, namespace="feature-flag")
@@ -24,7 +26,7 @@ app = FastAPI()
 async def create_flag(
     flag_data: dict, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    feature_flag = service.create_feature_flag(flag_data)
+    feature_flag = await service.create_feature_flag(flag_data)
     return jsonable_encoder(feature_flag)
 
 
@@ -32,7 +34,7 @@ async def create_flag(
 async def get_flag(
     id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    feature_flag = service.get_feature_flag(id)
+    feature_flag = await service.get_feature_flag(id)
     return jsonable_encoder(feature_flag)
 
 
@@ -42,22 +44,22 @@ async def update_flag(
     flag_data: dict,
     service: FeatureFlagService = Depends(get_feature_flag_service),
 ):
-    return service.update_feature_flag(id, flag_data)
+    return await service.update_feature_flag(id, flag_data)
 
 
 @app.delete("/api/feature-flags/{id}")
 async def delete_flag(
     id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    return service.delete_feature_flag(id)
+    return await service.delete_feature_flag(id)
 
 
 @app.post("/api/feature-flags/{flag_id}/enable", response_model=FeatureFlag)
 async def enable_feature_flag(
     flag_id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    service.enable_feature_flag(flag_id)
-    flag = service.get_feature_flag(flag_id)
+    await service.enable_feature_flag(flag_id)
+    flag = await service.get_feature_flag(flag_id)
     return flag
 
 
@@ -65,6 +67,6 @@ async def enable_feature_flag(
 async def disable_feature_flag(
     flag_id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    service.disable_feature_flag(flag_id)
-    flag = service.get_feature_flag(flag_id)
+    await service.disable_feature_flag(flag_id)
+    flag = await service.get_feature_flag(flag_id)
     return flag